% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R
\name{add_label}
\alias{add_label}
\title{Create a html variable for use as popups in interactive maps created with leaflet}
\usage{
add_label(data, label)
}
\arguments{
\item{data}{a data frame containing longitude and latitude}

\item{label}{a variable name of the data frame}
}
\value{
a tibble with html format label that works well with leaflet
}
\description{
Create a html variable for use as popups in interactive maps created with leaflet
}
