% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_points.R
\name{binning_points}
\alias{binning_points}
\title{Bin irregular simple feature into grids and weight the variables in each grid}
\usage{
binning_points(
  sf_object = data.points,
  cell_size = 10000,
  band_width = 20000,
  min_obs = 20,
  epsg_code = 32737,
  crs_code = 4326
)
}
\arguments{
\item{sf_object}{a simple feature of irregular points}

\item{cell_size}{the grid resolution in meters}

\item{band_width}{the band radius of searching for gridding in meters}

\item{min_obs}{minimum observation for smoothing should be integer}

\item{epsg_code}{the Universal Transverse Mercator code example 32737}

\item{crs_code}{the Geographical Coordinate System code example 4326}
}
\value{
a simple feature of polygons
}
\description{
Bin irregular simple feature into grids and weight the variables in each grid
}
