% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_tibble.R
\name{cnv_tibble}
\alias{cnv_tibble}
\title{Convert CTD profile data to tibble}
\usage{
cnv_tibble(cnv = cnv, interval = 2)
}
\arguments{
\item{cnv}{a scientific format of cnv after being converted from hex}

\item{interval}{depth interval to create standard depth}
}
\value{
a tibble with temperature, depth conductivity as standard CTD variables and additional variables depending on the sensors configured in the instrument
}
\description{
Convert CTD profile data to tibble
}
\examples{

}
