% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_corr.R
\name{compute_corr}
\alias{compute_corr}
\title{Compute correlation of two independent variables}
\usage{
compute_corr(data = cars, var1 = speed, var2 = dist)
}
\arguments{
\item{data}{input tibble or data frame}

\item{var1}{name of the first variable}

\item{var2}{name of the second variable}
}
\value{
a tibble with correlation coefficient and the p-value
}
\description{
Compute correlation of two independent variables
}
