% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{degree2fahrenheit}
\alias{degree2fahrenheit}
\title{Convert temperature from Degree to Fahrenheit}
\usage{
degree2fahrenheit(C_temp)
}
\arguments{
\item{C_temp}{Temperature in degree Celsius}
}
\value{

}
\description{
Convert temperature from Degree to Fahrenheit
}
