% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enso.R
\name{enso}
\alias{enso}
\title{Download Southern Oscillation Index Data}
\usage{
enso(begin_date, end_date)
}
\arguments{
\item{begin_date}{a begin date value you wish to download the data. Should be in a R format ("YYYY-MM-DD")}

\item{end_date}{an end date value you wish to download the data. Should be in a R format ("YYYY-MM-DD")}
}
\value{
a tibble contains;
\itemize{
\item date a date variable,
\item condition whether is warm or cool condition of the oceans,
\item phase whether is El-Nino or La-Nina and
\item soi variables
}
}
\description{
The Southern Oscillation Index (SOI) is defined as the standardized difference between barometric pressure between Tahiti and Darwin, Australia. The SOI is one measure of the large-scale fluctuations in air pressure occurring between the western and eastern tropical Pacific (i.e., the state of the Southern Oscillation) during El Niño and La Niña episodes. In general, smoothed time series of the SOI correspond very well with changes in ocean temperatures across the eastern tropical Pacific.
}
\examples{
\dontrun{
require(wior)
enso(begin_date = "2000-08-01", end_date = "2020-05-01")

}
}
\author{
Masumbuko Semba
}
