% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{fahrenheit2degree}
\alias{fahrenheit2degree}
\title{Convert temperature from Fahrenheit to Degree}
\usage{
fahrenheit2degree(F_temp)
}
\arguments{
\item{F_temp}{Temperature in fahrenheit}
}
\value{

}
\description{
Convert temperature from Fahrenheit to Degree
}
