% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_cap.R
\name{first_cap}
\alias{first_cap}
\title{Capitalizing first letter}
\usage{
first_cap(x)
}
\arguments{
\item{x}{a vector containing string or strings}
}
\value{
a vector of character(s) with each word begin with a captial letter
}
\description{
Capitalizing first letter
}
