% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1daysla.R
\name{get_1daysla}
\alias{get_1daysla}
\title{Get Sea level anomalies (SLA) from Altimetry at 25 spatial resolution near the Equator}
\usage{
get_1daysla(lon.min, lon.max, lat.min, lat.max, t1, t2)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 2017-02-13}

\item{t2}{a character end time in R-format ("year-month-day"). last date is often one day behind today.}
}
\value{
a tibble of daily sla within a specified geographical area and time limit.
}
\description{
Get Sea level anomalies (SLA) from Altimetry at 25 spatial resolution near the Equator
}
