% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chlSeawif.R
\name{get_chlSeawif}
\alias{get_chlSeawif}
\title{Download SeaWIFS chlorophyll}
\usage{
get_chlSeawif(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 1997-09-04}

\item{t2}{a character end time in R-format ("year-month-day"). The maximum data end on 210-12-11.}

\item{level}{represent mean average, where 1 = daily, 2 = 8 day compositie, 3 = monthly composite. Default is daily (level = 1)}
}
\value{
a tibble
}
\description{
The function help to download gridded data of chlorophyll-a at 9 kilometer resolution around the world. The data is retrieved from the ERDDAP server and you can obtain data at any location on the the major oceans and lakes
}
\examples{
\dontrun{
require(wior)
chl = get_chlSeawif(lon.min = 38.5,lon.max = 40, lat.min = -6, lat.max = -4,t1 = "2008-02-28", t2 = "2008-03-30", level = 2)
}

}
\author{
Masumbuko Semba
}
