% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meteo.R
\name{get_meteo}
\alias{get_meteo}
\title{Download hourly meteorological data}
\usage{
get_meteo(coastal_codes = 2, begin_year, end_year)
}
\arguments{
\item{coastal_codes}{a numeric number that represent a station for meteorological data (i.e 1 = Zanzibar, 2 = Tanga, 3 = Mwalimu Nyerere Airport, 4 = Pemba, 5 = Kilwa Masoko and 6 = Mtwara)}

\item{begin_year}{specify the minimal year you wish to extract the data}

\item{end_year}{specify the maximum year of the data you want to extract}
}
\value{
a tibble of meteorological data
}
\description{
Access surface meteorological data from the NOAA Integrated Surface Database from six stations along the coastal of Tanzania. These stations include Tanga, Pemba, Unguja, Mwalimu Nyerere International Airport, Kilwa Masoko and Mtwara.
}
\examples{
\dontrun{
require(wior)
site = get_meteo(coastal_codes = 3, begin_year = 2018, end_year = 2019)
}

}
\seealso{
\code{\link{get_sstMODIS}} and \code{\link{get_chlSeawif}}
}
\author{
Masumbuko Semba
}
