% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mld.R
\name{get_mld}
\alias{get_mld}
\title{Get climatological gridded mixed layer depth (MLD) from WOA}
\usage{
get_mld(lon.min, lon.max, lat.min, lat.max, t1 = NULL, t2 = NULL)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{not required for climatology}

\item{t2}{not required for climatology}
}
\value{
a tibble of daily MLD in a chosen geographical area.
}
\description{
Get climatological gridded mixed layer depth (MLD) from WOA
}
