% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oscar.R
\name{get_oscar}
\alias{get_oscar}
\title{Get OSCAR global surface currents at 1/3 degree spatial resolution at 5 day interval}
\usage{
get_oscar(lon.min, lon.max, lat.min, lat.max, t1, t2)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude valu}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 2002-10-01}

\item{t2}{a character end time in R-format ("year-month-day").It end one day before today date. Available until 2020-03-01}
}
\value{
a tibble of daily surface current (u and V) in specified geographical area and within the time limit.
}
\description{
Get OSCAR global surface currents at 1/3 degree spatial resolution at 5 day interval
}
