% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppMODIS.R
\name{get_ppMODIS}
\alias{get_ppMODIS}
\title{Download derived primary productivity from ERDDAP server}
\usage{
get_ppMODIS(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 2003-01-01}

\item{t2}{a character end time in R-format ("year-month-day"). The maximum data end two days before today.}

\item{level}{represent mean average, where 1 = daily, 2 = 3 day, 3 = monthly composite. Default is daily (level = 1)}
}
\value{
a tibble
}
\description{
Download derived primary productivity from ERDDAP server
}
