% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precip.R
\name{get_precip}
\alias{get_precip}
\title{Download precipation value at global scale with 5 kilometer resolution from Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS) since 1981}
\usage{
get_precip(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 1981-01-01}

\item{t2}{a character end time in R-format ("year-month-day"). The maximum data, when this function was written was 2020-01-01.}

\item{level}{represent mean average precipitation, where 1 = daily, 2 =  5-day composite, 3 = monthly composite, 4 = Annual composite. Default is daily (level = 1)}
}
\value{
a tibble
}
\description{
Download precipation value at global scale with 5 kilometer resolution from Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS) since 1981
}
\examples{
\dontrun{
require(tidyverse)
require(wior)
pangani = wior::get_precip(lon.min = 38.8, lon.max = 39.01, lat.min = -5.45, lat.max = -5.37, t1 = "1981-01-01", t2 = "2019-12-31", level = 4)
pangani \%>\% group_by(time) \%>\% summarise(precip = mean(precip, na.rm = TRUE)) \%>\% ggplot(aes(x = time, y = precip))+geom_path()
}

}
