% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainLand.R
\name{get_rainLand}
\alias{get_rainLand}
\title{Download rainfall average on land at 6 km resolution}
\usage{
get_rainLand(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 1981-01-01}

\item{t2}{a character end time in R-format ("year-month-day"). The maximum data end on 2019-10-01.}

\item{level}{represent mean average, where 1 = daily, 2 = 5 day, 3 = monthly, and 4 = annual average of precipitation. Default is daily (level = 1)}
}
\value{
a tibble
}
\description{
Download rainfall average on land at 6 km resolution
}
