% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sss.R
\name{get_sss}
\alias{get_sss}
\title{Get sea surface salinity (SSS) at 0.25 degree spatial resolution at 1 day interval}
\usage{
get_sss(lon.min, lon.max, lat.min, lat.max, t1, t2)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 2010-06-01}

\item{t2}{a character end time in R-format ("year-month-day").It end one day before today date. Available until two days before today}
}
\value{
a tibble of daily surface surface salinity in specified geographical area and within the time limit.
}
\description{
Get sea surface salinity (SSS) at 0.25 degree spatial resolution at 1 day interval
}
