% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sstMODIS.R
\name{get_sstMODIS}
\alias{get_sstMODIS}
\title{Get SST at global level for daily, 8 days and monthly composite from MODIS satellite data at 9 kilometer spatial resolution near the Equator}
\usage{
get_sstMODIS(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 2003-01-16}

\item{t2}{a character end time in R-format ("year-month-day"). last date is on 2919-10-16.}

\item{level}{dataset of SST for day (level = 1)  for 8 day (level = 2) and monthyl composite (level = 3)}
}
\value{
a tibble dataset contain the gridded sst
}
\description{
Get SST at global level for daily, 8 days and monthly composite from MODIS satellite data at 9 kilometer spatial resolution near the Equator
}
