% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sstMUR.R
\name{get_sstMUR}
\alias{get_sstMUR}
\title{Download Multi-scale Ultra-high Resolution SST at 1 km resolution}
\usage{
get_sstMUR(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 2002-06-01}

\item{t2}{a character end time in R-format ("year-month-day"). The maximum data end a day before today.}

\item{level}{represent mean average, where 1 = daily, 2 =  monthly composite, 3 = daily anomaly, 4 = monthly anomaly and 5 = climate. Default is daily (level = 1)}
}
\value{
a tibble
}
\description{
Download Multi-scale Ultra-high Resolution SST at 1 km resolution
}
