% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_windQuikscat.R
\name{get_windQuikscat}
\alias{get_windQuikscat}
\title{Get daily gloabal wind speed and stress from QuikSCAT satellite data at 12.5 kilometer spatial resolution near the Equator}
\usage{
get_windQuikscat(lon.min, lon.max, lat.min, lat.max, t1, t2, level = 1)
}
\arguments{
\item{lon.min}{a numeric minimum longitude value}

\item{lon.max}{a numeric maximum longitude value}

\item{lat.min}{a numeric minimum latitude value}

\item{lat.max}{a numeric maximum latitude value}

\item{t1}{a character begin time in R-format ("year-month-day"). The minimum begin from 1999-07-21}

\item{t2}{a character end time in R-format ("year-month-day"). last date is on 2009-11-21.}

\item{level}{whether is wind speed (level = 1) o  wind stress (level = 2)}
}
\value{
a tibble dataset contain the zonal and meridional field components
}
\description{
Get daily gloabal wind speed and stress from QuikSCAT satellite data at 12.5 kilometer spatial resolution near the Equator
}
