% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_2d.R
\name{interp_2d}
\alias{interp_2d}
\title{Interpolate irregular data to regular grids.}
\usage{
interp_2d(x, y, z, n)
}
\arguments{
\item{x}{A vector containing a element represent x value}

\item{y}{A vector of y-value}

\item{z}{A vector containing element of z-value}

\item{n}{determine the spacing, if you specify 10 you get 100 i.e 10*10 = 100}
}
\value{
A tibble format that is compatible with tidyverse ecosystem
}
\description{
Interpolate irregular data to regular grids.
}
