% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_hyperbolic.R
\name{inverse_hyperbolic}
\alias{inverse_hyperbolic}
\title{Transform data close to log() but allows for zeros}
\usage{
inverse_hyperbolic(x)
}
\arguments{
\item{x}{a numeric vector, a numeric variables in data frame or numeric matrix}
}
\value{
a numeric for vector, a data frame for data frame and matrix for matrix
}
\description{
Transform data close to log() but allows for zeros
}
