% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_symetry.R
\name{measure_location}
\alias{measure_location}
\title{Compute the central location values}
\usage{
measure_location(x)
}
\arguments{
\item{x}{a data frame or tibble with with continous variable(s)}
}
\value{
a atibble
}
\description{
This function compute the central values that help to describe the shape of distribution for continuous variable(s)
}
\examples{
\dontrun{
require(wior)
data = read_csv("data/meteo_coastal_station.csv")
data \%>\% measure_location()
}
}
\author{
Masumbuko Semba
}
