% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_symetry.R
\name{measure_symetry}
\alias{measure_symetry}
\title{Assess the shape of data}
\usage{
measure_symetry(x)
}
\arguments{
\item{x}{a data frame or tibble with with continous variable(s)}
}
\value{
a table containing a column of variable(s) and kurtosis and skewness for each variable
}
\description{
The function compute kurtosis and skewness to assess the shape of the distribution of the continuous values in the data frmae.
}
\examples{
\dontrun{
require(wior)
data = read_csv("data/meteo_coastal_station.csv")
data \%>\% measure_symetry()
}
}
\author{
Masumbuko Semba
}
