% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_symetry.R
\name{measure_variation}
\alias{measure_variation}
\title{Measure of variation}
\usage{
measure_variation(x)
}
\arguments{
\item{x}{a data frame or tibble with with continuous variable(s)}
}
\value{
a data frame with metrics of measure of variation  for each variable
}
\description{
This function compute the measure of variation from the data frame
}
\examples{
\dontrun{
require(wior)
data = read_csv("data/meteo_coastal_station.csv")
data \%>\% measure_variation()
}

}
\author{
Masumbuko Semba
}
