% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monsoon_season.R
\name{monsoon_season}
\alias{monsoon_season}
\title{Title:  A function that add a monsoon season column in the data frame from date}
\usage{
monsoon_season(data, x)
}
\arguments{
\item{data}{a Data frame}

\item{x}{a time variable in the data frame}
}
\value{
a tibble with added column of season
}
\description{
The Indian Ocean experience reversal of trade winds, which also affect the physical, biological and chemical process in the area. When you have a data frame with time variable, you can easily derive a colum with variables that shows the three common season---northeast (October--March), southeast (May--Septemeber) and inter--monsoon season (April and October). So with \code{monsoon_season}, we can easily convert any date and time into season
}
\examples{
\dontrun{
require(wior)
met = read_csv("data/met")
check the structure of the dataset
met \%>\% str()
# now we compute the monsoon
met \%>\% wior::monsoon_season(data = met, x = date)

}

}
\author{
Masumbuko semba
}
