% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_remove.R
\name{outlier_remove}
\alias{outlier_remove}
\title{Remove observations of a variables detected as outlier in a data frame}
\usage{
outlier_remove(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{a variable to remove outlier}
}
\value{
A tibble with only observations after removing rows of outliers
}
\description{
outliers are values that deviate so much from other observations.
}
