% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot vertical profile of CTD}
\usage{
plot_profile(data, x, y)
}
\arguments{
\item{data}{data frame}

\item{x}{a variable in data frame in the x-axis}

\item{y}{a variable in data frame in the y-axis}
}
\value{
a vertical profile plot of the selected variables
}
\description{
Plot vertical profile of CTD
}
