% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_tb.R
\name{raster_tb}
\alias{raster_tb}
\title{Convert raster file to tibble}
\usage{
raster_tb(path)
}
\arguments{
\item{path}{a character vector of full path name of the raster file}
}
\value{
A tibble of longitude, latitude and a z value
}
\description{
Convert raster file to tibble
}
