% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_crop.R
\name{sf_crop}
\alias{sf_crop}
\title{Crop point simple feature that are within a polygon boundary}
\usage{
sf_crop(sf_feature, polygon_sf)
}
\arguments{
\item{sf_feature}{a simple feature (point, line or polygon)}

\item{polygon_sf}{a polygon simple feature to crop point_sf}
}
\value{
A point simple feature of only point within the polygon
}
\description{
Crop point simple feature that are within a polygon boundary
}
