% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_tibble.R
\name{sf_tibble}
\alias{sf_tibble}
\title{Convert binned simple feature into tibble for drawing with ggplot2}
\usage{
sf_tibble(x)
}
\arguments{
\item{x}{a simple feature}
}
\value{
a data frame with three column of x,y, and z
}
\description{
Convert binned simple feature into tibble for drawing with ggplot2
}
