% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_sf.R
\name{tibble_sf}
\alias{tibble_sf}
\title{Convert table of longitude and latitude values into simple feature}
\usage{
tibble_sf(data, lon, lat)
}
\arguments{
\item{data}{data frame}

\item{lon}{longitude values}

\item{lat}{latitude values}
}
\value{
a simple feature with WGS84
}
\description{
Convert table of longitude and latitude values into simple feature
}
