% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\name{transect}
\alias{transect}
\title{Create transect from vector of longitude and latitude}
\usage{
transect(lon, lat, n)
}
\arguments{
\item{lon}{a vector with minimum and maximum value of longitude}

\item{lat}{a vector with minimum and maximum value of latitude}

\item{n}{the number of points representing quadrants along the transect}
}
\value{

}
\description{
Create transect from vector of longitude and latitude
}
