% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tb.R
\name{point_tb}
\alias{point_tb}
\title{Convert simple point feature into a tibble format based on original coordinate system (UTM or GCS)}
\usage{
point_tb(x)
}
\arguments{
\item{A}{point simple feature.}
}
\value{
It return a tibble containing the longitude and latitude and all the variables present in the sf object. If the original sf object was in Universal Transverse Mercator (UTM), you get the coordinates in meters and if in World Geodetic Cooridante (WGS), you get coordinate in degree. So you need to transform your simple feature in advance before based on which output you intend to get
}
\description{
Convert simple point feature into a tibble format based on original coordinate system (UTM or GCS)
}
